requirements.txt
requests>=2.28
pyyaml>=6.0
pandas>=2.0
openai>=1.0
